feat(genre): admin config

+++ src/genre/genre.service.ts
@@ -1,5 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { User } from 'src/user/entities/user.entity';
 import { handleError } from 'src/utils/handle-error.util';
 import { CreateGenreDto } from './dto/create-genre.dto';
 import { UpdateGenreDto } from './dto/update-genre.dto';
@@ -26,25 +31,47 @@ export class GenreService {
     return this.findById(id);
   }
 
-  create(dto: CreateGenreDto): Promise<Genre> {
-    const data: Genre = { ...dto };
-
-    return this.prisma.genre.create({ data }).catch(handleError);
+  async create(dto: CreateGenreDto, user: User): Promise<Genre> {
+    if (user.isAdmin) {
+      const genre: Genre = { ...dto };
+      return await this.prisma.genre
+        .create({
+          data: genre,
+        })
+        .catch(handleError);
+    } else {
+      throw new UnauthorizedException(
+        'You are not authorized to perform this action',
+      );
+    }
   }
 
-  async update(id: string, dto: UpdateGenreDto): Promise<Genre> {
-    await this.findById(id);
-    const data: Partial<Genre> = { ...dto };
+  async update(id: string, dto: UpdateGenreDto, user: User): Promise<Genre> {
+    if (user.isAdmin) {
+      await this.findById(id);
+      const data: Partial<Genre> = { ...dto };
 
-    return this.prisma.genre
-      .update({
-        where: { id },
-        data,
-      })
-      .catch(handleError);
+      return this.prisma.genre
+        .update({
+          where: { id },
+          data,
+        })
+        .catch(handleError);
+    } else {
+      throw new UnauthorizedException(
+        'You are not authorized to perform this action',
+      );
+    }
   }
-  async delete(id: string) {
-    await this.findById(id);
-    await this.prisma.genre.delete({ where: { id } });
+
+  async delete(id: string, user: User) {
+    if (user.isAdmin) {
+      await this.findById(id);
+      await this.prisma.genre.delete({ where: { id } });
+    } else {
+      throw new UnauthorizedException(
+        'You are not authorized to perform this action',
+      );
+    }
   }
 }
